requirements.txt
pytest==7.4.2
pytest-mock==3.11.1
python-dotenv==1.0.0
